Enable Legacy Record Download
The "old" machine, which is the black and white attendance machine uses
a slightly different algorithm for downloading records.
To allow backward compatibility, the additional line for downloading the
record from the machine in concern is also included.
Note that all legacy machine must have a machine id lower than 100, this
is specifically required for this application